backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
openai==1.58.1
anthropic==0.42.0
tavily-python==0.5.0
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
